balance game

+++ Assets/Scripts/Game.Runtime/Data/SpawnData.cs
@@ -5,7 +5,7 @@ namespace Game.Runtime
     [CreateAssetMenu(fileName = "SpawnData", menuName = "Data/SpawnData")]
     public class SpawnData : ScriptableObject
     {
-        public int timeCountDownSpawnEnemy;
+        public float timeCountDownSpawnEnemy;
         public EnemyBase[] enemyBases;
     }
 }
